requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.5
openai==0.28.0
requests==2.31.0
Pillow==10.0.0
google-api-python-client==2.100.0
python-dateutil==2.8.2
httpx==0.25.0
